pyproject.toml
[tool]
[tool.poetry]
authors = ["Miguel <miguel@aredpanic.com>"]
name = "root"
version = "0.0.0"
description = ''
[tool.poetry.dependencies]
dice = "^3.0"
"discord.py" = "*"
flask = "*"
jsonpickle = "^1.2"
python = "^3.8"
